juntar screen view model e edit model e acabar visual sets
