+++ src/app/components/screens/Train/Exercises/ProgrammedStretchingExercise.tsx
@@ -1,40 +1,47 @@
 import React, { useEffect, useState } from "react"
-import { StyleSheet } from "react-native"
+import { StyleSheet, View } from "react-native"
+import { IconButton } from "react-native-paper"
 import { RFValue } from "react-native-responsive-fontsize"
 import {
-  StretchingSetType,
-  StretchingSet as Set,
+  StretchingSetClass,
   Exercise
 } from "../../../../../dataDefinition/data"
-import { Theme } from "../../../../providers/Theme"
-import { Button } from "../../../reusable/Button"
+import { useTheme } from "../../../../providers/Theme"
 import InlineContainer from "../../../reusable/InlineContainer"
 import { Text } from "../../../reusable/Text"
+import { modelModes } from "../Models/Model"
 import StretchingSet from "./Sets/StretchingSet"
 
 export default function ProgrammedStretchingExercise({
-  theme,
   exNum,
   exercise,
-  isTS,
+  mode,
   onSetChange
 }: {
-  theme: Theme;
   exNum: number;
   exercise: Exercise;
-  onSetChange: (exNum: number, sets: StretchingSetType[]) => void;
-  isTS: boolean;
+  onSetChange: (exNum: number, sets: StretchingSetClass[]) => void;
+  mode: modelModes;
 }) {
-  const [sets, setSets] = useState<StretchingSetType[]>([])
+  const theme = useTheme()
+  const [sets, setSets] = useState<StretchingSetClass[]>([])
 
   useEffect(() => {
-    setSets(exercise.sets as StretchingSetType[])
+    setSets(exercise.sets as StretchingSetClass[])
   }, [])
 
   useEffect(() => {
     onSetChange(exNum, sets)
     /* console.log(`(PRE) SETS EXERCICIO ${exNum} -> ${JSON.stringify(sets)} `) */
   }, [sets])
+
+  function onSetCheckBox(setIndex: number) {
+    setSets((prevSets) => {
+      return prevSets.map((set, index) =>
+        index == setIndex ? { ...set, done: !set.done } : set
+      )
+    })
+  }
 
   function onWeightChange(setIndex: number, weight: number) {
     setSets((prevSets) => {
@@ -62,23 +69,20 @@ export default function ProgrammedStretchingExercise({
 
   function deleteSet(setIndex: number) {
     setSets((prevSets) => {
-      console.log(
-        `DELETING SET N ${setIndex}:  ${JSON.stringify(prevSets[setIndex])}`
-      )
       return prevSets.filter((set, index) => index != setIndex)
     })
   }
   function addSet() {
-    setSets((prevSets) => [...prevSets, new Set()])
+    setSets((prevSets) => [...prevSets, new StretchingSetClass()])
   }
 
-  const setElements = (exercise.sets as StretchingSetType[]).map(
+  const setElements = (exercise.sets as StretchingSetClass[]).map(
     (set, index) => (
       <StretchingSet
-        theme={theme}
-        model={isTS}
+        mode={mode}
         setNum={index}
         key={index}
+        done={set.done}
         weight={set.weight}
         wantedDuration={set.wantedDuration}
         duration={set.duration}
@@ -86,6 +90,7 @@ export default function ProgrammedStretchingExercise({
         onChangeWDuration={onChangeWDuration}
         onChangeDuration={onChangeDuration}
         onDeletePress={deleteSet}
+        onSetCheckbox={onSetCheckBox}
       />
     )
   )
@@ -130,28 +135,28 @@ export default function ProgrammedStretchingExercise({
         >
           Time
         </Text>
-        <Text
+        <View
           style={{
             ...styles.del,
-            ...styles.subtitle,
-            color: theme.colors.text
+            ...styles.subtitle
           }}
-        >
-          {""}
-        </Text>
+        />
       </InlineContainer>
 
       {setElements}
 
-      <Button
-        style={{
-          marginTop: theme.margins.m,
-          marginBottom: theme.margins.s
-        }}
-        onPress={addSet}
-      >
-        Add Set
-      </Button>
+      {mode == modelModes.Edit && (
+        <IconButton
+          color={theme.colors.primary}
+          size={30}
+          style={{
+            width: "100%",
+            alignSelf: "center"
+          }}
+          onPress={addSet}
+          icon={"plus"}
+        />
+      )}
     </>
   )
 }
