+++ src/app/components/screens/Train/Exercises/Sets/WESet.tsx
@@ -1,39 +1,45 @@
 import React from "react"
-import { StyleSheet } from "react-native"
+import { StyleSheet, View } from "react-native"
+import { Checkbox, IconButton } from "react-native-paper"
 import { RFValue } from "react-native-responsive-fontsize"
-import { Theme } from "../../../../../providers/Theme"
+import { images } from "../../../../../lib/extra"
+import { useTheme } from "../../../../../providers/Theme"
 import InlineContainer from "../../../../reusable/InlineContainer"
 import { Text } from "../../../../reusable/Text"
+import { modelModes } from "../../Models/Model"
 import SetFieldInput from "./SetFieldInput"
 
 export default function WESet({
-  theme,
-  model,
+  mode,
   setNum,
   weight,
   min,
   max,
+  done,
   onChangeWeight,
   onChangeRepMin,
   onChangeRepMax,
-  onDeletePress
+  onDeletePress,
+  onSetCheckbox
 }: {
-  theme: Theme;
-  model: boolean;
+  mode: modelModes;
   setNum: number;
   weight: number;
   min: number;
   max: number;
+  done: boolean;
   onChangeWeight: (setIndex: number, weight: number) => void;
   onChangeRepMin: (setIndex: number, amount: number) => void;
   onChangeRepMax: (setIndex: number, amount: number) => void;
   onDeletePress: (setIndex: number) => void;
+  onSetCheckbox: (setIndex: number) => void;
 }) {
+  const theme = useTheme()
   return (
     <InlineContainer
       style={{
         ...styles.container,
-        paddingVertical: theme.margins.s
+        paddingVertical: theme.margins.xs
       }}
     >
       <Text
@@ -47,24 +53,39 @@ export default function WESet({
         {setNum + 1}
       </Text>
       <SetFieldInput
+        inputMode={mode}
         style={styles.weight}
         value={isNaN(weight) ? "0" : weight.toString()}
         onChangeText={(w) => onChangeWeight(setNum, parseFloat(w))}
       />
       <SetFieldInput
+        inputMode={mode}
         style={styles.repRange}
-        disabled={model}
         value={isNaN(min) ? "0" : min.toString()}
         onChangeText={(n) => onChangeRepMin(setNum, parseInt(n))}
       />
       <SetFieldInput
+        inputMode={mode}
         style={styles.repRange}
-        disabled={model}
         value={isNaN(max) ? "0" : max.toString()}
         onChangeText={(n) => onChangeRepMax(setNum, parseInt(n))}
       />
-      <SetFieldInput style={styles.reps} disabled={!model} />
-      <IconButton icon={images.Trash} onPress={() => onDeletePress(setNum)} />
+      <View style={{ width: "10%", alignItems: "center" }}>
+        {mode == modelModes.Edit ? (
+          <IconButton
+            style={{ backgroundColor: theme.colors.primary, borderRadius: 5 }}
+            icon={images.Trash}
+            onPress={() => onDeletePress(setNum)}
+          />
+        ) : (
+          mode == modelModes.Session && (
+            <Checkbox
+              status={done ? "checked" : "unchecked"}
+              onPress={() => onSetCheckbox(setNum)}
+            />
+          )
+        )}
+      </View>
     </InlineContainer>
   )
 }
