+++ src/app/components/screens/Train/Models/ModelNav.tsx
@@ -3,13 +3,11 @@ import React from "react"
 import { Exercise } from "../../../../../dataDefinition/data"
 import ExerciseSelector from "../Exercises/ExerciseSelector"
 import { Exercise as ExerciseScreen } from "../Exercises/Exercise"
-import EditModel, { modelModes } from "./EditModel"
-import Model from "./Model"
+import Model, { modelModes } from "./Model"
 import ModelList, { TrainingModelDoc } from "./ModelList"
 
 export type RootStackParamListModelNav = {
   ModelList: undefined;
-  Model: { model: TrainingModelDoc };
   EditModel: { model?: TrainingModelDoc; mode: modelModes };
   ExerciseSelector: { onSubmit: (selectedExercises: Exercise[]) => void };
   Exercise: { exercise: Exercise };
@@ -24,8 +22,7 @@ export default function ExerciseNav() {
       screenOptions={{ headerShown: false }}
     >
       <Stack.Screen name="ModelList" component={ModelList} />
-      <Stack.Screen name="Model" component={Model} />
-      <Stack.Screen name="EditModel" component={EditModel} />
+      <Stack.Screen name="EditModel" component={Model} />
       <Stack.Screen name="ExerciseSelector" component={ExerciseSelector} />
       <Stack.Screen name="Exercise" component={ExerciseScreen} />
     </Stack.Navigator>
